frontend: creating a condition to check availability of the experts
backend: creating api to create a case

+++ SOS-server/database/migrations/2022_07_28_090814_create_cases_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->integer('user_id');
             $table->integer('expert_id');
             $table->integer('is_done');
-            $table->integer('user_lat');
-            $table->integer('user_long');
+            $table->float('user_lat', 9, 7);
+            $table->float('user_long', 9, 7);
             $table->timestamps();
         });
     }
